pyproject.toml
[tool.poetry]
name = "bed-plam"
version = "0.1.0"
description = ""
authors = ["anhelina <123456765432za@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
djangorestframework = "^3.14.0"
markdown = "^3.5.2"
django-cors-headers = "^4.3.1"
google-api-python-client = "^2.117.0"
nltk = "^3.8.1"
tika = "^2.6.0"
langdetect = "^1.0.9"
stanza = "^1.7.0"
spacy = "^3.7.4"
fake-headers = "^1.0.2"
aiohttp = "^3.9.5"
openai = "^1.27.0"
scikit-learn = "^1.4.2"
transformers = "^4.40.2"
torch = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
